refactor: `WithdrawRequest` 모델의 위치 변경 및 관련 로직 수정
- `WithdrawRequest` 모델을 `model/member` 패키지에서 `network/request` 패키지로 이동했습니다.
- `MemberRepository`의 `withdraw` 함수 시그니처를 변경했습니다.
    - 기존: `fun withdraw(request: WithdrawRequest): Flow<Unit>`
    - 변경: `fun withdraw(reason: WithdrawReason, customReason: String? = null): Flow<Unit>`
